pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = "A simple Telegram bot"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
flake8-docstrings = "^1.7.0"
flake8-import-order = "^0.18.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
